requirements.txt
django
celery
redis
boto3
gitpython
djangorestframework
djoser
djangorestframework-simplejwt
gunicorn
